hot-fix in monitor util imports, cleaned up
